update handling of translated titles
see https://kjc-sv013.kjc.uni-heidelberg.de/redmine/issues/2368:

Take the Japanese entry as an example: http://east.uni-hd.de/buddh/ind/28/92/1182/
Should we put the translation of the title of a Japanese journal after its transliteration?

We now have the form: Indian Logic // Indo Ronrigaku Kenkyū]
My suggestion: Indo Ronrigaku Kenkyū [Indian Logic] (This is also the form in which we represent the title of a Japanese article)
Add

Updated by Patrick Mc Allister 22 days ago

    Subject changed from Transliteration and Translation to Transliteration and Translation of Japanese journal titles
    Status changed from New to In Progress
    Assignee set to Patrick Mc Allister

Comment Edit

I agree this should be the correct form. Currently Tamboti is down, however, so I can't do much about it at the moment.
Updated by Birgit Kellner 22 days ago
Comment Edit

According to the guidelines (Dropbox): "For titles of articles and books, give the English translation in square brackets after the romanized Japanese title. If no English translation is available in the publication itself, please produce one yourself, and add an asterisk before the English translation." I have amended this to include journal titles. So Pei-Lin's suggestion is correct, but one has to check whether there is an English-language title in the publication itself, or whether it is provided by us.

"Indian logic" is not a full translation of "Indo Ronrigaku Kenkyū", which rather translates as "Studies in/of Indian Logic". But "Indian logic" may be the official English-language title given in the publicaiton. Please check this with Watanabe-san.
Updated by Birgit Kellner 8 days ago
Comment Edit

Recently updated entries by Chiou still contain the "//"-solution, which is against our guidelines.

In the future, please enquire in unclear cases BEFORE updating the entries, otherwise we get chaos. Updates by editors should always be consistent with the guidelines. If there is uncertainty in individual cases, find solutions, and only update after the solution is settled and has become a part of the guidelines.

Please correct these entries as soon as possible, or, if there are still open questions, post them here. Otherwise, please close this bug so that we know you have followed this discussion, and that the issue is resolved.